requirements.txt
Django==4.1.3
Pillow==9.3.0
django-ckeditor==6.5.1